SpringBoot_微信点餐系统/sell/src/main/java/com/imooc/service/impl/SellerServiceImpl.java
package com.imooc.service.impl;

import com.imooc.dateobject.SellerInfo;
import com.imooc.repository.SellerInfoRepository;
import com.imooc.service.SellerService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.util.StringUtils;

/**
 * 卖家端
 * @author kenshin
 * @date 2018/8/2 下午3:26
 */
@Service
@Slf4j
public class SellerServiceImpl implements SellerService {

    @Autowired
    private SellerInfoRepository sellerInfoRepository;

    @Override
    public SellerInfo findSellerInfoByOpenId(String openid) {
        if (StringUtils.isEmpty(openid)){
            log.error("openid传空给我 你查毛线啊!");
            return null;
        }
        SellerInfo sellerInfo = sellerInfoRepository.findByOpenid(openid);
        return sellerInfo;

    }
}
